network-intelligence/python-services/crewai-api/requirements.txt
# CrewAI Multi-Agent Framework
crewai[tools]==0.86.0
crewai-tools==0.17.0

# FastAPI Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20

# HTTP Client for calling Next.js services
httpx==0.28.1
aiohttp==3.11.10

# LLM Integration
openai==1.57.2
anthropic==0.40.0
langchain-openai>=0.2.1
langchain-anthropic>=0.3.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
